🚚(frontend) consolidate under the `index` filename convention
We had a mix of `moduleName/moduleName.ts(x)` and `moduleName/index.ts(x)`
in our files. This was not ideal but not a big deal as long as all files
were strictly internal to the project.

Now that we're working to enable module overriding, we decided to unify
all paths under a single convention. It seemed cleaner to us to use
the `index` convention as it avoids having a redundancy in basically
every module path in the project.
