+++ src/frontend/js/components/CourseGlimpse/index.spec.tsx
@@ -4,7 +4,7 @@ import { render } from '@testing-library/react';
 import React from 'react';
 import { IntlProvider } from 'react-intl';
 
-import { CourseGlimpse } from './CourseGlimpse';
+import { CourseGlimpse } from '.';
 
 describe('components/CourseGlimpse', () => {
   const course = {
