+++ src/frontend/js/components/CourseGlimpse/index.tsx
