+++ src/frontend/js/components/CourseGlimpseList/index.spec.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import { IntlProvider } from 'react-intl';
 
 import { Course } from 'types/Course';
-import { CourseGlimpseList } from './CourseGlimpseList';
+import { CourseGlimpseList } from '.';
 
 describe('components/CourseGlimpseList', () => {
   it('renders a list of Courses into a list of CourseGlimpses', () => {
