+++ src/frontend/js/components/PaginateCourseSearch/index.spec.tsx
@@ -4,7 +4,7 @@ import { fireEvent, render } from '@testing-library/react';
 import React from 'react';
 import { IntlProvider } from 'react-intl';
 
-import { CourseSearchParamsContext } from 'data/useCourseSearchParams/useCourseSearchParams';
+import { CourseSearchParamsContext } from 'data/useCourseSearchParams';
 import { PaginateCourseSearch } from '.';
 
 describe('<PaginateCourseSearch />', () => {
