+++ src/frontend/js/components/Search/_styles.scss
