+++ src/frontend/js/components/SearchFilterGroup/_styles.scss
