+++ src/frontend/js/components/SearchFilterGroupModal/_styles.scss
