+++ src/frontend/js/components/SearchFilterGroupModal/index.tsx
@@ -6,8 +6,8 @@ import {
 } from 'react-intl';
 import ReactModal from 'react-modal';
 
-import { fetchList } from 'data/getResourceList/getResourceList';
-import { CourseSearchParamsContext } from 'data/useCourseSearchParams/useCourseSearchParams';
+import { fetchList } from 'data/getResourceList';
+import { CourseSearchParamsContext } from 'data/useCourseSearchParams';
 import { requestStatus } from 'types/api';
 import { FilterDefinition, FilterValue } from 'types/filters';
 import { Nullable } from 'utils/types';
