+++ src/frontend/js/components/SearchFilterValueParent/index.spec.tsx
@@ -4,15 +4,15 @@ import { fireEvent, render, wait } from '@testing-library/react';
 import React from 'react';
 import { IntlProvider } from 'react-intl';
 
-jest.mock('data/getResourceList/getResourceList', () => ({
+jest.mock('data/getResourceList', () => ({
   fetchList: jest.fn(),
 }));
 
-import { fetchList } from 'data/getResourceList/getResourceList';
-import { CourseSearchParamsContext } from 'data/useCourseSearchParams/useCourseSearchParams';
+import { fetchList } from 'data/getResourceList';
+import { CourseSearchParamsContext } from 'data/useCourseSearchParams';
 import { APIListRequestParams } from 'types/api';
 import { jestMockOf } from 'utils/types';
-import { SearchFilterValueParent } from './SearchFilterValueParent';
+import { SearchFilterValueParent } from '.';
 
 const mockFetchList: jestMockOf<typeof fetchList> = fetchList as any;
