+++ src/frontend/js/components/SearchFiltersPane/_styles.scss
