+++ src/frontend/js/components/SearchLoader/_styles.scss
