+++ src/frontend/js/components/SearchLoader/index.tsx
