+++ src/frontend/js/data/getResourceList/index.ts
