+++ src/frontend/js/data/useCourseSearch/index.ts
@@ -3,10 +3,7 @@ import { useState } from 'react';
 import { APIListRequestParams } from 'types/api';
 import { Nullable } from 'utils/types';
 import { useAsyncEffect } from 'utils/useAsyncEffect';
-import {
-  fetchList,
-  fetchListResponse,
-} from '../getResourceList/getResourceList';
+import { fetchList, fetchListResponse } from '../getResourceList';
 
 export const useCourseSearch = (searchParams: APIListRequestParams) => {
   const [courseSearchResponse, setCourseSearchResponse] = useState<
