+++ src/frontend/js/data/useCourseSearchParams/index.ts
